[XLA:GPU][IndexAnalysis] Do not remove unused RTVars during simplification.
When RTVar is produced by HLOConstant, then it gets removed during RTVars optimization. The problem is that if there are two or more RTVars after the constant one, then they are not adjusted by shifting their SymbolIDs.

It is easier to remove this logic completely and rely on IndexingMap::RemoveUnusedSymbols instead.

PiperOrigin-RevId: 666733253